backend/faceIDVerification/requirements.txt
azure-functions
requests
azure-cognitiveservices-vision-face
